// src/application/interfaces/presenters/dto/question/create.question.dto.ts
export type CreateQuestionDTO = {
    title: string;
    description: string;
    content: string;
    tags: string[];
    totalPotentialMarks: number;
    difficultyLevel: number;
    type: 'multiple_choice' | 'true_or_false';
    isTrue?: boolean;
    multipleChoiceOptions?: {
        content: string;
        isCorrect: boolean;
    }[]
}